separate file for RenderbufferAttachment
